added artefact configurations

+++ wso2is-setup/startup.sh
@@ -52,11 +52,17 @@ if [ -f "$scriptFilelst" ] && [ -s "$scriptFilelst" ]; then
     esac
 
     if [ "$crr_md5" != "$remt_md5" ]; then
-        printf "\nUpdate found for the script, hence updating."
+    
+        read -p 'NEW UPDATE FOUND ! Do you wish to update the script [yes]: ' updateScript
+        updateScript=$(echo "$updateScript" | awk '{print tolower($0)}')
+        if [ "$updateScript" == "yes" ] || [ "$updateScript" == "y" ]; then
+         printf "\nUpdating the script file .."
         mv $scriptFilelst $scriptFile
         chmod 755 $scriptFile
         printf "\nPlease run it again !!\n"
         exit
+        fi
+
     else
         rm -rf $scriptFilelst
     fi
@@ -78,9 +84,16 @@ isVersionIndex=$(expr ${#is_versions_arr[@]} - 1)
 dbType="h2"
 startupPrams=""
 
+if [ ! -f $script_dir/.artefacts.tar ]; then
+    curl -sfL https://github.com/dmxunlimit/tools/raw/master/wso2is-setup/.artefacts.tar -o $script_dir/.artefacts.tar
+    mkdir $script_dir/artefacts
+    tar -xf $script_dir/.artefacts.tar -C $script_dir/artefacts
+fi
 
-curl -sfL https://github.com/dmxunlimit/tools/raw/master/wso2is-setup/.artefacts.tar -o $script_dir/.artefacts.tar
-tar -xf $script_dir/.artefacts.tar -C $script_dir/artefacts
+if [ ! -d $script_dir/artefacts ]; then
+    mkdir $script_dir/artefacts
+    tar -xf $script_dir/.artefacts.tar -C $script_dir/artefacts
+fi
 
 dockerStart() {
